ledger: spending_reqs: Move check_input_lock to spending_reqs.
